+++ src/models/analytics-panel.ts
@@ -59,6 +59,7 @@ const schema = Joi.object().keys({
   layout: layoutSchema.default([]),
   cards: Joi.array().items(Joi.object().keys({
     title: Joi.string().required().example('My widget'),
+    subtitle: Joi.string().allow('').example('My widget description'),
     query: analyticsQuerySchema.required(),
     visualisation: analyticsVisualisationSchema.required(),
   })).required(),
@@ -81,6 +82,7 @@ interface AnalyticsPanel {
   layout: AnalyticsPanelLayout;
   cards: {
     title: string;
+    subtitle?: string;
     query: AnalyticsQuery;
     visualisation: AnalyticsVisualisation;
   }[];