+++ src/models/command-type.ts
@@ -2,14 +2,14 @@ import Joi from 'joi';
 
 import { schema as baseFieldConfigurationSchema, BaseFieldConfiguration } from './fields/base-field-configuration';
 
-const schema = Joi.object().keys({
+const schema = (apiVersion: number): Joi.ObjectSchema => Joi.object().keys({
   hashId: Joi.string().required().example('x18a92'),
   name: Joi.string().required().example('Measurement cycle'),
   start: Joi.string().valid('required', 'optional', 'disabled').required().example('required')
     .description('\'required\': user must provide command.startAt. \'optional\': user can provide command.startAt or a delay for the command to start after it is sent to the device. \'disabled\': user cannot provide command.startAt nor a delay.'),
   end: Joi.string().valid('required', 'optional', 'disabled').required().example('disabled')
     .description('\'required\': user must provide command.endAt. \'optional\': user can provide command.endAt. \'disabled\': user cannot provide command.endAt.'),
-  fieldConfigurations: Joi.array().items(baseFieldConfigurationSchema).required()
+  fieldConfigurations: Joi.array().items(baseFieldConfigurationSchema(apiVersion)).required()
     .description('See the chapter on open fields on how to use this'),
   channelSelect: Joi.string().valid('single', 'multiple', 'off').required().example('off')
     .description('When creating a command of this type, the user can then optionally choose one (in case of \'single\') or more channelIndices (in case of \'multiple\') for which this command is relevant. If \'off\' is chosen, the user cannot specify channelIndices'),