+++ src/models/device-type.ts
@@ -1,16 +1,18 @@
 import Joi from 'joi';
 import { schema as baseFieldConfigurationSchema, BaseFieldConfiguration } from './fields/base-field-configuration';
 
-const baseSchema = Joi.object().keys({
+const schema = (apiVersion: number): Joi.ObjectSchema => Joi.object().keys({
   hashId: Joi.string().required().example('wasd2'),
   name: Joi.string().required().example('Cathodic protection device'),
-  fieldConfigurations: Joi.array().items(baseFieldConfigurationSchema).required()
+  fieldConfigurations: Joi.array().items(baseFieldConfigurationSchema(apiVersion)).required()
     .description('See the chapter on open fields on how to use this'),
-  pinGroupFieldConfigurations: Joi.array().items(baseFieldConfigurationSchema).required()
+  pinGroupFieldConfigurations: Joi.array()
+    .items(baseFieldConfigurationSchema(apiVersion))
+    .required()
     .description('Defines deviceFields on the location (pinGroup) the device is connected to. Can be used in report type functions. See the chapter on open fields on how to use this'),
   channels: Joi.array().items(Joi.object().keys({
     name: Joi.string().required().example('Red wire'),
-    pinFieldConfigurations: Joi.array().items(baseFieldConfigurationSchema).required()
+    pinFieldConfigurations: Joi.array().items(baseFieldConfigurationSchema(apiVersion)).required()
       .description('Defines deviceFields on the pin the channel is connected to. Can be used in report type functions. See the chapter on open fields on how to use this'),
     defaultPinName: Joi.string().example('Anode').description('If undefined, the channel cannot be linked to a pin'),
     charts: Joi.array().items(Joi.object().keys({
@@ -32,10 +34,7 @@ const baseSchema = Joi.object().keys({
     })).required(),
   })).required(),
   commandTypeHashIds: Joi.array().items(Joi.string().example('x18a92')).required().description('The hashIds of the command types a user can schedule for this device'),
-});
-
-const schema = baseSchema
-  .tag('deviceType')
+}).tag('deviceType')
   .description('Information about the type of device');
 
 interface DeviceType {
@@ -67,5 +66,5 @@ interface DeviceType {
 }
 
 export {
-  schema, baseSchema, DeviceType,
+  schema, DeviceType,
 };