+++ src/models/supplier-report-type.ts
@@ -2,20 +2,17 @@ import Joi from 'joi';
 
 import { schema as baseFieldConfigurationSchema, BaseFieldConfiguration } from './fields/base-field-configuration';
 
-const baseSchema = Joi.object().keys({
+const schema = (apiVersion: number): Joi.ObjectSchema => Joi.object().keys({
   hashId: Joi.string().required().example('l19a7s'),
   name: Joi.string().required().example('Temperature and inclination'),
   fieldConfigurations: Joi.object().keys({
-    pinGroup: Joi.array().items(baseFieldConfigurationSchema).required(),
-    pin: Joi.array().items(baseFieldConfigurationSchema).required(),
-    measurement: Joi.array().items(baseFieldConfigurationSchema).required(),
+    pinGroup: Joi.array().items(baseFieldConfigurationSchema(apiVersion)).required(),
+    pin: Joi.array().items(baseFieldConfigurationSchema(apiVersion)).required(),
+    measurement: Joi.array().items(baseFieldConfigurationSchema(apiVersion)).required(),
   }).required()
     .description('See the chapter on open fields on how to use this'),
   deletedAt: Joi.date().allow(null).required().example(null),
-});
-
-const schema = baseSchema
-  .description('An object defining what a condition report should look like, used in connectivity environments')
+}).description('An object defining what a condition report should look like, used in connectivity environments')
   .tag('supplierReportType');
 
 interface SupplierReportType {
@@ -29,4 +26,4 @@ interface SupplierReportType {
   deletedAt: Date | null;
 }
 
-export { schema, baseSchema, SupplierReportType };
+export { schema, SupplierReportType };