+++ src/routes/device/link.ts
@@ -3,7 +3,6 @@ import { ControllerGeneratorOptionsWithClient } from '../../comms/controller';
 
 import { schema as deviceSchema, Device } from '../../models/device';
 import { schema as deviceTypeSchema, DeviceType } from '../../models/device-type';
-import { schema as pinSchema, Pin } from '../../models/pin';
 import { schema as fieldsToServerUpdateSchema, FieldsToServerUpdate } from '../../models/fields/fields-to-server-update';
 
 interface Request {
@@ -24,7 +23,6 @@ interface Response {
   device: Device;
   deviceType: DeviceType;
   nextReportBefore: Date | null;
-  pins: Pin[];
 }
 
 const controllerGeneratorOptions: ControllerGeneratorOptionsWithClient = {
@@ -41,12 +39,11 @@ const controllerGeneratorOptions: ControllerGeneratorOptionsWithClient = {
       deviceFields: fieldsToServerUpdateSchema,
     })).required(),
   }).required(),
-  response: Joi.object().keys({
+  response: (apiVersion: number): Joi.ObjectSchema => Joi.object().keys({
     device: deviceSchema.required(),
-    deviceType: deviceTypeSchema.required(),
+    deviceType: deviceTypeSchema(apiVersion).required(),
     measurementCycle: Joi.valid(null).default(null), // for legacy purposes only
     nextReportBefore: Joi.date().allow(null).example('2019-12-31T15:25Z').required(),
-    pins: Joi.array().items(pinSchema).required().description('All ports (pins) of the location (pinGroup), as some might have an updated deviceFields property'),
   }).required(),
   right: { environment: 'SENSORS' },
   description: 'Connect a device to a pin group (and its channels to the pin group\'s pins. Future condition reports from this device will also be available on this pin group. A claim token for this device will be invalidated',