+++ src/routes/index.ts
@@ -19,6 +19,7 @@ import ReportRoute from './report';
 import ReportTypeRoute from './report-type';
 import SettingsRoute from './settings';
 import SubscriptionRoute from './subscription';
+import SuggestionsRoute from './suggestions';
 import SupplierRoute from './supplier';
 import SupplierActivityRoute from './supplier-activity';
 import SupplierCertificateRoute from './supplier-certificate';
@@ -65,6 +66,8 @@ class Routes {
 
   subscription: SubscriptionRoute;
 
+  suggestions: SuggestionsRoute;
+
   supplier: SupplierRoute;
 
   supplierActivity: SupplierActivityRoute;
@@ -97,6 +100,7 @@ class Routes {
     this.reportType = new ReportTypeRoute(comms);
     this.settings = new SettingsRoute(comms);
     this.subscription = new SubscriptionRoute(comms);
+    this.suggestions = new SuggestionsRoute(comms);
     this.supplier = new SupplierRoute(comms);
     this.supplierActivity = new SupplierActivityRoute(comms);
     this.supplierCertificate = new SupplierCertificateRoute(comms);
@@ -127,6 +131,7 @@ export {
   ReportTypeRoute,
   SettingsRoute,
   SubscriptionRoute,
+  SuggestionsRoute,
   SupplierRoute,
   SupplierActivityRoute,
   SupplierCertificateRoute,