+++ src/routes/routes.ts
@@ -17,6 +17,7 @@ import * as report from './report/routes';
 import * as reportType from './report-type/routes';
 import * as settings from './settings/routes';
 import * as subscription from './subscription/routes';
+import * as suggestions from './suggestions/routes';
 import * as supplier from './supplier/routes';
 import * as supplierActivity from './supplier-activity/routes';
 import * as supplierCertificate from './supplier-certificate/routes';
@@ -44,6 +45,7 @@ export {
   reportType,
   settings,
   subscription,
+  suggestions,
   supplier,
   supplierActivity,
   supplierCertificate,