+++ src/routes/supplier-report-type/update.ts
@@ -29,12 +29,12 @@ const controllerGeneratorOptions: ControllerGeneratorOptionsWithSupplier = {
   params: Joi.object().keys({
     hashId: Joi.string().required().example('y124as'),
   }).required(),
-  body: Joi.object().keys({
+  body: (apiVersion: number): Joi.ObjectSchema => Joi.object().keys({
     name: Joi.string().example('Temperature'),
     fieldConfigurations: Joi.object().keys({
-      pinGroup: updatableFieldConfigurationsSchema,
-      pin: updatableFieldConfigurationsSchema,
-      measurement: updatableFieldConfigurationsSchema,
+      pinGroup: updatableFieldConfigurationsSchema(apiVersion),
+      pin: updatableFieldConfigurationsSchema(apiVersion),
+      measurement: updatableFieldConfigurationsSchema(apiVersion),
     }).description('See the chapter on open fields on how to use this'),
     parser: Joi.string().example('[omitted]').description('A javascript function that parses an incoming report. See the chapter "User defined code"'),
     retryFailedSince: Joi.date().description('If supplied, all incoming reports of this type that failed to decode since the provided date will be reparsed.'),